% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_generation.R
\name{add_cohort}
\alias{add_cohort}
\title{Add age cohorts to survey data}
\usage{
add_cohort(data, ageq)
}
\arguments{
\item{data}{Survey data imported/read into R in dataframe format}

\item{ageq}{Question number in survey that asks for respondent age}
}
\value{
A new "\code{Cohort}" column in the dataframe
}
\description{
Add age cohorts to survey data
}
\examples{
\dontrun{
svy |> add_cohort("Q3")
}
}
