% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_table.R
\name{build_table}
\alias{build_table}
\title{Build results tables}
\usage{
build_table(tbl, ci)
}
\arguments{
\item{tbl}{Dataframe outputted from the \code{summarize_question()} function. If crosstabbed results, the \code{summarize_question()} function outputs two dataframes in a named list.}

\item{ci}{Confidence interval for significant testing}
}
\value{
DT datatable (really most useful in the app)
}
\description{
Table builds depend on analysis type. If single variable, it's a straightforward table. If crosstab, two tables are created with significance testing - results and sigtest within each group, and results and sigtest between groups.
}
\examples{
\dontrun{
build_table(tbl$out_bwn, .95)
}
}
