% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{convertMenuItem}
\alias{convertMenuItem}
\title{Convert one menu item to another in shiny}
\usage{
convertMenuItem(mi, tabName)
}
\arguments{
\item{mi}{blah}

\item{tabName}{blah}
}
\value{
blah
}
\description{
Convert one menu item to another in shiny
}
\keyword{internal}
