% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_labelled}
\alias{create_labelled}
\title{Convert user generated variables into labelled data (\code{haven} labelled spss data)}
\usage{
create_labelled(tbl, toc)
}
\arguments{
\item{tbl}{Survey}

\item{toc}{TOC}
}
\value{
Survey data frame with labelled user generated vars
}
\description{
Convert user generated variables into labelled data (\code{haven} labelled spss data)
}
\examples{
\dontrun{
create_labelled(tbl, toc)
}
}
