% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_sids.R
\name{fetch_sids}
\alias{fetch_sids}
\title{Fetch all survey IDs from Qualtrics}
\usage{
fetch_sids()
}
\value{
Dataframe with survey dates, survey IDs and survey names
}
\description{
This function assumes that the Qualtrics API key and subdomain are in the .renviron
}
\examples{
\dontrun{
sids <- fetch_sids()
}
}
