% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{filter_choices}
\alias{filter_choices}
\title{Filtered choices available to user in app}
\usage{
filter_choices(tbl, var)
}
\arguments{
\item{tbl}{Survey data}

\item{var}{Labels of specific variable}
}
\value{
named vector
}
\description{
Filtered choices available to user in app
}
\keyword{internal}
