% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_responses.R
\name{get_responses}
\alias{get_responses}
\title{Get survey responses}
\usage{
get_responses(
  data,
  group,
  gsub,
  var,
  qname,
  qsub,
  selector,
  resps,
  filters,
  ci
)
}
\arguments{
\item{data}{Survey data}

\item{group}{Grouping variable if used}

\item{gsub}{Group sublabel if applicable}

\item{var}{Survey variable of interest}

\item{qname}{Question name of variable}

\item{qsub}{Any sub-label or extra definition pulled from the API}

\item{selector}{Selector type - if Likert the app will allow for top/bottom box}

\item{resps}{Vector of all possible questions respondents could answer}

\item{filters}{List containing any Q columns to filter and the responses to keep}

\item{ci}{Confidence interval used for stat sig testing}
}
\value{
Data frame with with all grouped data and the survey proportion
}
\description{
This is a helper function that is called by the \code{summarize_survey()} function.
As a result, all parameters are passed into it via the higher order function.
}
\examples{
\dontrun{
na
}
}
