% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likert_box.R
\name{likert_box}
\alias{likert_box}
\title{Calculate top and bottom box}
\usage{
likert_box(tbl, top, bottom)
}
\arguments{
\item{tbl}{Dataframe summary returned by the \code{summarize_survey()} function}

\item{top}{Vector of numbers for the top; the dataframe has variable numbers which can be used}

\item{bottom}{Vector of numbers for the top; the dataframe has variable numbers which can be used}
}
\value{
Dataframe with summed proportion and proper labels; CIs are dropped
}
\description{
After summarizing the survey question, a user can then choose to summarize further into Top/Bottom box. Rather than re-running the analysis, this function just summarizes the table further. As a result, confidence intervals are lost.
}
\examples{
\dontrun{
top <- c(3,4); bottom <- c(1,2)
likert_box(tbl, top, bottom)
}
}
