% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pull_filter_toc}
\alias{pull_filter_toc}
\title{Build filter options seen in app dropdown}
\usage{
pull_filter_toc(toc, blockfilter)
}
\arguments{
\item{toc}{TOC}

\item{blockfilter}{Filter}
}
\value{
updated TOC
}
\description{
Build filter options seen in app dropdown
}
\keyword{internal}
