% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{remove_html}
\alias{remove_html}
\title{Helper function - remove html}
\usage{
remove_html(string)
}
\arguments{
\item{string}{blah}
}
\value{
string
}
\description{
Helper function - remove html
}
\keyword{internal}
