% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/significance_test.R
\name{significance_test}
\alias{significance_test}
\title{Test statistical significance}
\usage{
significance_test(tbl, conf_level)
}
\arguments{
\item{tbl}{Dataframe with summarized results}

\item{conf_level}{Confidence interval e.g., .9, .95}
}
\value{
Dataframe with appended significance letters. Letters are appended to the variable labels as well to match the test results
}
\description{
If it's a single variable (not crosstabbed) each selection option is compared against all others via the \code{prop.test()} function. If it is crosstabbed, selection options within each group are compared.
}
\examples{
\dontrun{
significance_test(tbl, conf_level = .95)
}
}
