% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sublabel_clean.R
\name{subquestion_clean}
\alias{subquestion_clean}
\title{Clean the sub-label of the question}
\usage{
subquestion_clean(tbl, qsub)
}
\arguments{
\item{tbl}{Summarized survey question}

\item{qsub}{Question sub-labels passed through by the \code{summarize_question()} function}
}
\value{
Dataframe, but sub-labels are either NA or hopefully informative
}
\description{
This is a helper function used in the \code{summarize_survey()} function. It is used to map any sub-labels pulled from the question wordings against the variable labels pulled from the SPSS data. In the event that there is a match, variable labels win. If possible, we try to split the qsubs on a " - " and then try to match the split labels against the variable labels again. The goal is to try to get to meaningful qsubs for grouping in tables and graphs.
}
\examples{
\dontrun{
na
}
}
