% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_block.R
\name{summarize_block}
\alias{summarize_block}
\title{Summarize a full survey block from Qualtrics}
\usage{
summarize_block(block_choice, parameters, data)
}
\arguments{
\item{block_choice}{What is the survey block to be summarized}

\item{parameters}{What are the parameters to pass through for crosstabbing and confidence intervals}

\item{data}{The survey data to pass through}
}
\value{
List of data frames; each with the question number and whether the siginficance testing was within/between groups
}
\description{
Function will take the survey block and all other parameters (confidence interval, crosstabs, and any filters) and run analysis and significance testing
}
\examples{
\dontrun{
tbl <- summarize_block("Block2", parameters, data)
}
}
