% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_question.R
\name{summarize_question}
\alias{summarize_question}
\title{Summarize survey question of interest}
\usage{
summarize_question(parameters, data)
}
\arguments{
\item{parameters}{Parameters are a named list with the question ids for the grouping variable and variable of interest. The question ids are used to filter the table of contents to pull all potential columns of interest.}

\item{data}{Named list that is returned by the \code{fetch_survey_data()} function.}
}
\value{
Data frame with all proportions and confidence intervals for the variable of interest, split out by grouping variable (if one is selected).
}
\description{
This is the wrapper function that summarizes all survey data
}
\examples{
\dontrun{
parameters = list(qid = "VARID", gid = "GROUPID")
tbl <- summarize_survey(parameters, data = out)
}
}
