% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{survey_duration}
\alias{survey_duration}
\title{Calculate mean survey duration across respondents}
\usage{
survey_duration(x)
}
\arguments{
\item{x}{blah}
}
\value{
time variable
}
\description{
Calculate mean survey duration across respondents
}
\keyword{internal}
