% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl_augment.R
\name{tbl_augment}
\alias{tbl_augment}
\title{Augment a table with missing rows}
\usage{
tbl_augment(tbl)
}
\arguments{
\item{tbl}{A dataframe that was processed by the \code{survey_summary()} function}
}
\value{
Data frame
}
\description{
If a scale to a survey question has zero responses, it will be dropped from the analysis because the variables being passed into the \code{group_by()} function are SPSS <dbl+lbl> variable types, rather than factors. This is an absolute kludge of a solution to make sure the empty response rows are retained in the data.
}
\examples{
\dontrun{
tbl_augment(tbl)
}
}
