% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{toc_filter}
\alias{toc_filter}
\title{Filter table of contents by block}
\usage{
toc_filter(tbl, blck)
}
\arguments{
\item{tbl}{Survey}

\item{blck}{Survey block}
}
\value{
Filtered toc
}
\description{
Filter table of contents by block
}
\keyword{internal}
