% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{user_generated}
\alias{user_generated}
\title{Check to see if user generated variables are available in the survey}
\usage{
user_generated(tbl, toc)
}
\arguments{
\item{tbl}{Survey}

\item{toc}{Table of contents}
}
\value{
Updated table of contents
}
\description{
These variables must begin with an "x"
On survey load, the function find the user generated variables and adds them to the TOC
}
\examples{
\dontrun{
user_generated(tbl, toc)
}
}
