% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xlsx_build.R
\name{xlsx_build}
\alias{xlsx_build}
\title{Build xlsx table output}
\usage{
xlsx_build(data, file)
}
\arguments{
\item{data}{List of dataframes that have been processed via \code{summarize_block()}}

\item{file}{File name of outputted Excel file}
}
\value{
Formatted tables in Excel
}
\description{
Function to output Excel tables - one sheet per question within a survey block. The function gets called in the Shiny app within the download handler.
}
\examples{
\dontrun{
xlsx_build(list, "filename.xlsx")
}
}
