% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svy_summary.R
\name{svy_summary}
\alias{svy_summary}
\title{Summarize the survey data}
\usage{
svy_summary(tbl, block, group, gsub, var, qname, qsub, selector)
}
\arguments{
\item{tbl}{Survey data with group and variable of interest}

\item{block}{Survey block section}

\item{group}{Grouping variable (if used)}

\item{gsub}{Group sublabel if applicable}

\item{var}{Survey variable of interest}

\item{qname}{Question name pulled from the TOC}

\item{qsub}{qsub pulled from the TOC}

\item{selector}{Selector type - if Likert the app will allow for top/bottom box}
}
\value{
Data frame with all grouped analysis
}
\description{
This is a helper function for \code{summarize_survey()} and all arguments are passed into this via that higher order function.
}
\examples{
\dontrun{
na
}
}
